packages/adapters/pyproject.toml
[project]
name = "ky-adapters"
version = "0.1.0"
description = "ky-platform external data source adapters (KIS / DART / FRED / ECOS / KOSIS / EIA / EXIM)"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["ky_adapters*"]
